bugfix: create a base controller that requires authentication
and inherit it from all the controllers except public controllers that do
not require authentication.

fix specs so that protected contollers require authentication

fixes #47
