+++ app/controllers/dashboard_controller.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 # Dashboard for logged-in users
-class DashboardController < ApplicationController
-  before_action :authenticate_user!
-
+class DashboardController < ProtectedController
   def index; end
 end
