+++ app/controllers/protected_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# a base class for contollers that requires authentication
+class ProtectedController < ApplicationController
+  before_action :authenticate_user!
+end
