+++ app/controllers/public_controller.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# a base class for contollers that does not require authentication
+class PublicController < ApplicationController
+end
