+++ app/controllers/welcome_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # A controller for guest users
-class WelcomeController < ApplicationController
+class WelcomeController < PublicController
   def index
   end
 end
