+++ spec/requests/staff_users_controller_spec.rb
@@ -1,0 +1,145 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+# This spec was generated by rspec-rails when you ran the scaffold generator.
+# It demonstrates how one might use RSpec to test the controller code that
+# was generated by Rails when you ran the scaffold generator.
+#
+# It assumes that the implementation code is generated by the rails scaffold
+# generator. If you are using any extension libraries to generate different
+# controller code, this generated spec may or may not pass.
+#
+# It only uses APIs available in rails and/or rspec-rails. There are a number
+# of tools you can use to make these specs even more expressive, but we're
+# sticking to rails and rspec-rails APIs to keep things simple and stable.
+
+RSpec.describe StaffUsersController, type: :request do
+  it_behaves_like "authentication is required", Rails.application.routes.url_helpers.staff_users_path
+
+  context "when user is authenticated" do
+    login_user
+
+    # This should return the minimal set of attributes required to create a valid
+    # StaffUser. As you add validations to StaffUser, be sure to
+    # adjust the attributes here as well.
+    let(:valid_attributes) do
+      { name: "foo", display_name: "Foo", active: true, account: "foo12345" }
+    end
+
+    let(:invalid_attributes) do
+      { name: "foo", display_name: "Foo", active: true, account: "foo" }
+    end
+
+    describe "GET /index" do
+      it "renders a successful response" do
+        StaffUser.create! valid_attributes
+        get staff_users_url
+        expect(response).to be_successful
+      end
+
+      it_behaves_like "success response", Rails.application.routes.url_helpers.staff_users_url(only_path: true) do
+        before do
+          StaffUser.create! valid_attributes
+        end
+      end
+    end
+
+    describe "GET /show" do
+      it "renders a successful response" do
+        staff_user = StaffUser.create! valid_attributes
+        get staff_user_url(staff_user)
+        expect(response).to be_successful
+      end
+    end
+
+    describe "GET /new" do
+      it "renders a successful response" do
+        get new_staff_user_url
+        expect(response).to be_successful
+      end
+    end
+
+    describe "GET /edit" do
+      it "renders a successful response" do
+        staff_user = StaffUser.create! valid_attributes
+        get edit_staff_user_url(staff_user)
+        expect(response).to be_successful
+      end
+    end
+
+    describe "POST /create" do
+      context "with valid parameters" do
+        it "creates a new StaffUser" do
+          expect do
+            post staff_users_url, params: { staff_user: valid_attributes }
+          end.to change(StaffUser, :count).by(1)
+        end
+
+        it "redirects to the created staff_user" do
+          post staff_users_url, params: { staff_user: valid_attributes }
+          expect(response).to redirect_to(staff_user_url(StaffUser.last))
+        end
+      end
+
+      context "with invalid parameters" do
+        it "does not create a new StaffUser" do
+          expect do
+            post staff_users_url, params: { staff_user: invalid_attributes }
+          end.not_to change(StaffUser, :count)
+        end
+
+        it "renders a response with 422 status (i.e. to display the 'new' template)" do
+          post staff_users_url, params: { staff_user: invalid_attributes }
+          expect(response).to have_http_status(:unprocessable_entity)
+        end
+      end
+    end
+
+    describe "PATCH /update" do
+      context "with valid parameters" do
+        let(:new_account) { "foo123456789" }
+        let(:new_attributes) do
+          { name: "foo", display_name: "Foo", active: true, account: new_account }
+        end
+
+        it "updates the requested staff_user" do
+          staff_user = StaffUser.create! valid_attributes
+          patch staff_user_url(staff_user), params: { staff_user: new_attributes }
+          staff_user.reload
+          expect(staff_user.account).to eq new_account
+        end
+
+        it "redirects to the staff_user" do
+          staff_user = StaffUser.create! valid_attributes
+          patch staff_user_url(staff_user), params: { staff_user: new_attributes }
+          staff_user.reload
+          expect(response).to redirect_to(staff_user_url(staff_user))
+        end
+      end
+
+      context "with invalid parameters" do
+        it "renders a response with 422 status (i.e. to display the 'edit' template)" do
+          staff_user = StaffUser.create! valid_attributes
+          patch staff_user_url(staff_user), params: { staff_user: invalid_attributes }
+          expect(response).to have_http_status(:unprocessable_entity)
+        end
+      end
+    end
+
+    describe "DELETE /destroy" do
+      it "destroys the requested staff_user" do
+        staff_user = StaffUser.create! valid_attributes
+        expect do
+          delete staff_user_url(staff_user)
+        end.to change(StaffUser, :count).by(-1)
+      end
+
+      it "redirects to the staff_users list" do
+        staff_user = StaffUser.create! valid_attributes
+        delete staff_user_url(staff_user)
+        expect(response).to redirect_to(staff_users_url)
+      end
+    end
+  end
+end
